training works

+++ MAD/ST_MLAD/utils.py
@@ -37,7 +37,8 @@ class Visualizer(object):
 def process_feat(feat, length):
     new_feat = np.zeros((length, feat.shape[1])).astype(np.float32)
     
-    r = np.linspace(0, len(feat), length+1, dtype=np.int)
+    #r = np.linspace(0, len(feat), length+1, dtype=np.int)
+    r = np.linspace(0, len(feat), length+1, dtype=np.int32)
     for i in range(length):
         if r[i]!=r[i+1]:
             new_feat[i,:] = np.mean(feat[r[i]:r[i+1],:], 0)
